+++ CHANGELOG.md
@@ -11,8 +11,11 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Fixed
 - run_program:  Added disconnect and error message if master and/or slave fails replication requirements check.
-- run_program:  Added reconnect due to long term dump/restores cause connection timeouts.
+- run_program:  Added connect_chk call due to long term dump/restores cause connection timeouts.
 - dump_load_dbs:  Fixed problem with mutable default arguments issue.
+
+### Added
+- connect_chk:  Checks to see if the connection is still active and reconnect if not.
 
 ### Changed
 - chk_rep_cfg:  Remove innodb_support_xa from replication config for MySQL 8.0 and above.